Added a new class for every enemy
